Add Applications V1 API endpoint (#316)
* Relax user passwords

In reality our users will be a handfull and privacy oriented. We do not
have to add these tight restrictions only for backoffice users.

* Add users to seeds

* Rework database

It is okay to play with these since we are not live yet. I have taken
the proper precautions on the deployment side to erase conflicting
tables.

* Add Application resources API route